+++ tools/module.go
@@ -119,6 +119,7 @@ func moduleRunE(cmd *cobra.Command, args []string) error {
 			module.GetKindStore().UpdatePolicy,
 			module.GetKindStore().ValueType,
 			stateStore,
+			nil,
 		)
 		cli.NoError(err, "unable to create store config")